refactor: move the TaskDisplay to other folder
